% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{swap_control_site_name}
\alias{swap_control_site_name}
\title{Control site names}
\usage{
swap_control_site_name(x, quiet = FALSE)
}
\arguments{
\item{x}{Character vector of site names.}

\item{quiet}{Logical. If \code{TRUE}, suppresses messages about replaced values.
Default: \code{FALSE}.}
}
\value{
\code{x} with elements matched to the thesaurus replaced by their canonical forms.

Unless \code{quiet = TRUE}, the function also prints a list of values that were
changed. A warning will be issued if elements of \code{x} were not matched to the
thesaurus (regardless of the \code{quiet} argument).
}
\description{
This function ensures site names conform to a controlled vocabulary.
It compares a vector of names to a thesaurus of variant names, misspellings
and character encoding inconsistencies and replaces matches with a canonical
form.
}
\details{
Elements of \code{x} are matched to the site name thesaurus using the following
rules (in order of priority):
\enumerate{
\item Exact matches to a variant name in the thesaurus
\item Matches if non-ASCII characters are substituted for another a character
(e.g. "Catalhoyuk" → "Çatalhöyük")
}

Ambiguous site names—elements of \code{x} that match multiple thesaurus entries—
will cause an error.
}
\examples{
sites <- c("Catalhoyuk", "Tepe Ganj Dareh", "Jericho")
swap_control_site_name(sites)
}
\concept{controlled vocabulary functions}
