% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapdata-package.R
\docType{package}
\name{swapdata-package}
\alias{swapdata}
\alias{swapdata-package}
\title{swapdata: Southwest Asian Palaeoarchaeology Data}
\description{
A collection of archaeological datasets and tools related to the 
    prehistory of Southwest Asia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/swapdata}
  \item Report bugs at \url{https://github.com/joeroe/swapdata/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
